Add macos unit tests, optimize code, refactor timer (#5796)

+++ src/server/reactor_thread.cc
@@ -936,7 +936,7 @@ void Server::join_heartbeat_thread() {
     if (heartbeat_thread.joinable()) {
         swoole_trace_log(SW_TRACE_SERVER, "terminate heartbeat thread");
         if (pthread_cancel(heartbeat_thread.native_handle()) > 0) {
-            swoole_sys_warning("pthread_cancel(%ld) failed", heartbeat_thread.native_handle());
+            swoole_sys_warning("pthread_cancel(%ld) failed", (long) heartbeat_thread.native_handle());
         }
         // wait thread
         heartbeat_thread.join();
@@ -966,7 +966,7 @@ void Server::join_reactor_thread() {
         } else {
         _cancel:
             if (pthread_cancel(thread->thread.native_handle()) > 0) {
-                swoole_sys_warning("pthread_cancel(%ld) failed", thread->thread.native_handle());
+                swoole_sys_warning("pthread_cancel(%ld) failed", (long) thread->thread.native_handle());
             }
         }
         thread->thread.join();
